pyproject.toml
[tool.poetry]
name = "nonebot-plugin-envious"
version = "0.0.1"
description = "羡慕 koishi"
authors = ["fllesser <fllessive@gmail.com>"]
readme = "README.md"
packages = [{include = "nonebot_plugin_envious"}]
repository = "https://github.com/fllesser/nonebot-plugin-envious"
keywords = ["nonebot", "nonebot2", "envious", "koishi" ,"羡慕"]

[tool.poetry.dependencies]
python = ">=3.10"
nonebot2 = ">=2.3.0"
nonebot-adapter-onebot = ">=2.4.6"
nonebot-plugin-localstore = ">=0.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"